requirements.txt
python-dotenv
pymysql-pooling
content_size_limit_asgi